CL-15: changed input value for addCredit to float instead of integer, also added rounding to bigdecimal in controller to ensure correct rounding of balance

+++ src/main/java/com/MIW/Cohort5/Clups/model/User.java
@@ -45,11 +45,11 @@ public class User {
 
     public User(){}
 
-    public void addToBalance(Integer amount) {
+    public void addToBalance(BigDecimal amount) {
         if (prepaidBalance == null){
-            prepaidBalance = BigDecimal.valueOf(amount);
+            prepaidBalance = amount;
         } else {
-            prepaidBalance = prepaidBalance.add(BigDecimal.valueOf(amount));
+            prepaidBalance = prepaidBalance.add(amount);
         }
     }
