+++ src/main/java/com/MIW/Cohort5/Clups/services/implementations/UserDetailsServiceImpl.java
@@ -227,7 +227,7 @@ public class UserDetailsServiceImpl implements UserService {
         return model;
     }
 
-    public void addCredit(Integer userCode, Integer amount) {
+    public void addCredit(Integer userCode, BigDecimal amount) {
         User user = userRepository.findUserByUserCode(userCode);
 
         user.addToBalance(amount);
